[CPU] ExperimentalDetectronROIFeatureExtractor: JIT kernel
