+++ src/plugins/intel_cpu/src/nodes/experimental_detectron_roifeatureextractor.cpp
@@ -18,62 +18,58 @@ namespace intel_cpu {
 namespace node {
 namespace {
 
-// implementation taken from Caffe2
-template <typename T>
-struct PreCalc {
-    int pos1;
-    int pos2;
-    int pos3;
-    int pos4;
-    T w1;
-    T w2;
-    T w3;
-    T w4;
+struct LevelParams {
+    const float *src;
+    int height;
+    int width;
 };
 
 template <typename T>
 void pre_calc_for_bilinear_interpolate(
+        const int n_roi,
         const int height,
         const int width,
-        const int pooled_height,
-        const int pooled_width,
-        const int iy_upper,
-        const int ix_upper,
-        T roi_start_h,
-        T roi_start_w,
-        T bin_size_h,
-        T bin_size_w,
-        int roi_bin_grid_h,
-        int roi_bin_grid_w,
-        std::vector<PreCalc<T>>& pre_calc) {
+        const int pooled_h,
+        const int pooled_w,
+        const int sampling_ratio_h,
+        const int sampling_ratio_w,
+        T y1,
+        T x1,
+        T bin_height,
+        T bin_width,
+        T sample_distance_h,
+        T sample_distance_w,
+        std::vector<std::vector<T>>& weights,
+        std::vector<std::vector<int>>& indices) {
+    constexpr size_t bli_params_num = 4u;
     int pre_calc_index = 0;
-    for (int ph = 0; ph < pooled_height; ph++) {
-        for (int pw = 0; pw < pooled_width; pw++) {
-            for (int iy = 0; iy < iy_upper; iy++) {
-                const T yy = roi_start_h + ph * bin_size_h +
-                             static_cast<T>(iy + .5f) * bin_size_h /
-                             static_cast<T>(roi_bin_grid_h);  // e.g., 0.5, 1.5
-                for (int ix = 0; ix < ix_upper; ix++) {
-                    const T xx = roi_start_w + pw * bin_size_w +
-                                 static_cast<T>(ix + .5f) * bin_size_w /
-                                 static_cast<T>(roi_bin_grid_w);
+
+    weights[n_roi].resize(bli_params_num * sampling_ratio_h * sampling_ratio_w * pooled_h * pooled_w);
+    indices[n_roi].resize(bli_params_num * sampling_ratio_h * sampling_ratio_w * pooled_h * pooled_w);
+
+    for (int ph = 0; ph < pooled_h; ph++) {
+        for (int pw = 0; pw < pooled_w; pw++) {
+            for (int iy = 0; iy < sampling_ratio_h; iy++) {
+                const T yy = y1 + ph * bin_height + static_cast<T>(iy + .5f) * sample_distance_h;  // e.g., 0.5, 1.5
+                for (int ix = 0; ix < sampling_ratio_w; ix++) {
+                    const T xx = x1 + pw * bin_width + static_cast<T>(ix + .5f) * sample_distance_w;
 
                     T x = xx;
                     T y = yy;
                     // deal with: inverse elements are out of feature map boundary
                     if (y < -1.0 || y > height || x < -1.0 || x > width) {
-                        // empty
-                        PreCalc<T> pc;
-                        pc.pos1 = 0;
-                        pc.pos2 = 0;
-                        pc.pos3 = 0;
-                        pc.pos4 = 0;
-                        pc.w1 = 0;
-                        pc.w2 = 0;
-                        pc.w3 = 0;
-                        pc.w4 = 0;
-                        pre_calc.at(pre_calc_index) = pc;
-                        pre_calc_index += 1;
+                        // weights and indices are 0
+                        weights[n_roi][pre_calc_index] = static_cast<T>(0);
+                        weights[n_roi][pre_calc_index + 1] = static_cast<T>(0);
+                        weights[n_roi][pre_calc_index + 2] = static_cast<T>(0);
+                        weights[n_roi][pre_calc_index + 3] = static_cast<T>(0);
+
+                        indices[n_roi][pre_calc_index] = 0;
+                        indices[n_roi][pre_calc_index + 1] = 0;
+                        indices[n_roi][pre_calc_index + 1] = 0;
+                        indices[n_roi][pre_calc_index + 1] = 0;
+
+                        pre_calc_index += bli_params_num;
                         continue;
                     }
 
@@ -91,14 +87,14 @@ void pre_calc_for_bilinear_interpolate(
 
                     if (y_low >= height - 1) {
                         y_high = y_low = height - 1;
-                        y = (T)y_low;
+                        y = static_cast<T>(y_low);
                     } else {
                         y_high = y_low + 1;
                     }
 
                     if (x_low >= width - 1) {
                         x_high = x_low = width - 1;
-                        x = (T)x_low;
+                        x = static_cast<T>(x_low);
                     } else {
                         x_high = x_low + 1;
                     }
@@ -106,205 +102,45 @@ void pre_calc_for_bilinear_interpolate(
                     T ly = y - y_low;
                     T lx = x - x_low;
                     T hy = static_cast<T>(1) - ly, hx = static_cast<T>(1) - lx;
-                    T w1 = hy * hx, w2 = hy * lx, w3 = ly * hx, w4 = ly * lx;
 
                     // save weights and indices
-                    PreCalc<T> pc;
-                    pc.pos1 = y_low * width + x_low;
-                    pc.pos2 = y_low * width + x_high;
-                    pc.pos3 = y_high * width + x_low;
-                    pc.pos4 = y_high * width + x_high;
-                    pc.w1 = w1;
-                    pc.w2 = w2;
-                    pc.w3 = w3;
-                    pc.w4 = w4;
-                    pre_calc[pre_calc_index] = pc;
+                    weights[n_roi][pre_calc_index] = hy * hx;
+                    weights[n_roi][pre_calc_index + 1] = hy * lx;
+                    weights[n_roi][pre_calc_index + 2] = ly * hx;
+                    weights[n_roi][pre_calc_index + 3] = ly * lx;
 
-                    pre_calc_index += 1;
+                    indices[n_roi][pre_calc_index] = y_low * width + x_low;
+                    indices[n_roi][pre_calc_index + 1] = y_low * width + x_high;
+                    indices[n_roi][pre_calc_index + 2] = y_high * width + x_low;
+                    indices[n_roi][pre_calc_index + 3] = y_high * width + x_high;
+
+                    pre_calc_index += bli_params_num;
                 }
             }
         }
     }
 }
 
-template <typename T>
-void ROIAlignForward_cpu_kernel(
-        const int nthreads,
-        const T* bottom_data,
-        const T& spatial_scale,
-        const int channels,
-        const int height,
-        const int width,
-        const int pooled_height,
-        const int pooled_width,
-        const int sampling_ratio,
-        const T* bottom_rois,
-        const bool aligned,
-        T* top_data) {
-    int roi_cols = 4;
+void roiTargetLevels(const float* rois, int* levels, const size_t num_rois, const int num_levels) {
+    constexpr float canonical_scale = 224.0f;
+    constexpr int canonical_level = 2;
 
-    int n_rois = nthreads / channels / pooled_width / pooled_height;
-    // (n, c, ph, pw) is an element in the pooled output
-    parallel_for(n_rois, [&](size_t n) {
-        int index_n = n * channels * pooled_width * pooled_height;
-
-        // roi could have 4 or 5 columns
-        const T* offset_bottom_rois = bottom_rois + n * roi_cols;
-        int roi_batch_ind = 0;
-        if (roi_cols == 5) {
-            roi_batch_ind = static_cast<int>(offset_bottom_rois[0]);
-            offset_bottom_rois++;
-        }
-
-        T offset = aligned ? (T)0.5 : (T)0.0;
-        // Do not using rounding; this implementation detail is critical
-        T roi_start_w = offset_bottom_rois[0] * spatial_scale - offset;
-        T roi_start_h = offset_bottom_rois[1] * spatial_scale - offset;
-        T roi_end_w = offset_bottom_rois[2] * spatial_scale - offset;
-        T roi_end_h = offset_bottom_rois[3] * spatial_scale - offset;
-
-        // Force malformed ROIs to be 1x1
-        T roi_width = (std::max)(roi_end_w - roi_start_w, (T)1.);
-        T roi_height = (std::max)(roi_end_h - roi_start_h, (T)1.);
-        T bin_size_h = static_cast<T>(roi_height) / static_cast<T>(pooled_height);
-        T bin_size_w = static_cast<T>(roi_width) / static_cast<T>(pooled_width);
-
-        // We use roi_bin_grid to sample the grid and mimic integral
-        int roi_bin_grid_h = (sampling_ratio > 0)
-                             ? sampling_ratio
-                             : static_cast<int>(ceil(roi_height / pooled_height));  // e.g., = 2
-        int roi_bin_grid_w =
-                (sampling_ratio > 0) ? sampling_ratio : static_cast<int>(ceil(roi_width / pooled_width));
-
-        // We do average (integral) pooling inside a bin
-        const T count = static_cast<T>(roi_bin_grid_h * roi_bin_grid_w);  // e.g. = 4
-
-        // we want to precalculate indices and weights shared by all chanels,
-        // this is the key point of optimiation
-        std::vector<PreCalc<T>> pre_calc(
-                roi_bin_grid_h * roi_bin_grid_w * pooled_width * pooled_height);
-        pre_calc_for_bilinear_interpolate(
-                height,
-                width,
-                pooled_height,
-                pooled_width,
-                roi_bin_grid_h,
-                roi_bin_grid_w,
-                roi_start_h,
-                roi_start_w,
-                bin_size_h,
-                bin_size_w,
-                roi_bin_grid_h,
-                roi_bin_grid_w,
-                pre_calc);
-
-        for (int c = 0; c < channels; c++) {
-            int index_n_c = index_n + c * pooled_width * pooled_height;
-            const T* offset_bottom_data =
-                    bottom_data + (roi_batch_ind * channels + c) * height * width;
-            int pre_calc_index = 0;
-
-            for (int ph = 0; ph < pooled_height; ph++) {
-                for (int pw = 0; pw < pooled_width; pw++) {
-                    int index = index_n_c + ph * pooled_width + pw;
-
-                    T output_val = 0.;
-                    for (int iy = 0; iy < roi_bin_grid_h; iy++) {
-                        for (int ix = 0; ix < roi_bin_grid_w; ix++) {
-                            PreCalc<T> pc = pre_calc[pre_calc_index];
-                            output_val += pc.w1 * offset_bottom_data[pc.pos1] +
-                                          pc.w2 * offset_bottom_data[pc.pos2] +
-                                          pc.w3 * offset_bottom_data[pc.pos3] +
-                                          pc.w4 * offset_bottom_data[pc.pos4];
-
-                            pre_calc_index += 1;
-                        }
-                    }
-                    output_val /= count;
-
-                    top_data[index] = output_val;
-                }  // for pw
-            }  // for ph
-        }  // for c
-    });
-}
-
-
-void redistribute_rois(const float* rois, int* level_ids,
-                       const int num_rois, const int levels_num) {
-    const float canonical_scale = 224.0f;
-    const int canonical_level = 2;
-
-    for (int i = 0; i < num_rois; ++i) {
+    for (size_t i = 0; i < num_rois; ++i) {
         const float x0 = rois[4 * i + 0];
         const float y0 = rois[4 * i + 1];
         const float x1 = rois[4 * i + 2];
         const float y1 = rois[4 * i + 3];
 
-        int target_level = levels_num;
+        int target_level = num_levels;
         float area = (x1 - x0) * (y1 - y0);
         if (area > 0) {
             area = std::sqrt(area) / canonical_scale;
             area = std::log2(area + 1e-6f);
             target_level = static_cast<int>(std::floor(area + canonical_level));
-            target_level = (std::max)(0, (std::min)(levels_num - 1, target_level));
+            target_level = std::max(0, std::min(num_levels - 1, target_level));
         }
 
-        level_ids[i] = target_level;
-    }
-}
-
-
-void reord(const float* src_data, const int* ranks, const int n, const int step, float* dst_data,
-             int* dst_mapping) {
-    std::iota(dst_mapping, dst_mapping + n, 0);
-    std::sort(dst_mapping, dst_mapping + n, [&ranks](size_t i1, size_t i2) {return ranks[i1] < ranks[i2];});
-    for (int i = 0; i < n; ++i) {
-        const int j = dst_mapping[i];
-        assert(0 <= j && j < n);
-        cpu_memcpy(dst_data + i * step, src_data + j * step, sizeof(float) * step);
-    }
-}
-
-void split_points(const std::vector<int>& ids, std::vector<int>& rois_per_level, const int levels_num) {
-    rois_per_level.clear();
-    rois_per_level.resize(levels_num, 0);
-    for (size_t i = 0; i < ids.size(); ++i) {
-        assert(0 <= ids[i] && ids[i] < levels_num);
-        rois_per_level[ids[i]]++;
-    }
-    for (int i = 1; i < levels_num; ++i) {
-        rois_per_level[i] += rois_per_level[i - 1];
-    }
-    rois_per_level.insert(rois_per_level.begin(), 0);
-}
-
-
-void reorder_rois(const float *rois, const int* ids, int* mapping, const int rois_num,
-                  float * reordered_rois, std::vector<int>& rois_per_level, const int levels_num) {
-    rois_per_level.clear();
-    rois_per_level.resize(levels_num, 0);
-    for (int i = 0; i < rois_num; ++i) {
-        assert(0 <= ids[i] && ids[i] < levels_num);
-        rois_per_level[ids[i]]++;
-    }
-    for (int i = 1; i < levels_num; ++i) {
-        rois_per_level[i] += rois_per_level[i - 1];
-    }
-    rois_per_level.insert(rois_per_level.begin(), 0);
-
-    std::vector<int> level_counter = rois_per_level;
-
-    for (int i = 0; i < rois_num; ++i) {
-        const int level = ids[i];
-        assert(level < levels_num);
-        const int j = level_counter[level];
-        assert(0 <= j && j < rois_num);
-        reordered_rois[j * 4 + 0] = rois[i * 4 + 0];
-        reordered_rois[j * 4 + 1] = rois[i * 4 + 1];
-        reordered_rois[j * 4 + 2] = rois[i * 4 + 2];
-        reordered_rois[j * 4 + 3] = rois[i * 4 + 3];
-        level_counter[level]++;
+        levels[i] = target_level;
     }
 }
 
@@ -357,11 +193,40 @@ void ExperimentalDetectronROIFeatureExtractor::initSupportedPrimitiveDescriptors
                          impl_desc_type::ref_any);
 }
 
+void ExperimentalDetectronROIFeatureExtractor::createPrimitive() {
+    Node::createPrimitive();
+
+    const bool jit_is_beneficial = (sampling_ratio_ >= 3 || sampling_ratio_ == 0);
+    if (!roi_align_kernel_ && jit_is_beneficial) {
+        createJitKernel(getParentEdgeAt(0)->getMemoryPtr()->getDesc().getPrecision(), ROIAlignLayoutType::ncsp); 
+    }
+}
+
+void ExperimentalDetectronROIFeatureExtractor::createJitKernel(const InferenceEngine::Precision& dataPrec, const ROIAlignLayoutType& selectLayout) {
+    auto jcp = jit_roi_align_params();
+    jcp.alg = Algorithm::ROIAlignAvg;
+    jcp.data_prc = dataPrec;
+    jcp.data_size = dataPrec.size();
+    jcp.layout = selectLayout;
+    jcp.pooled_h = pooled_height_;
+    jcp.pooled_w = pooled_width_;
+
+    if (mayiuse(cpu::x64::avx512_core)) {
+        roi_align_kernel_.reset(new jit_uni_roi_align_kernel_f32<cpu::x64::avx512_core>(jcp));
+    } else if (mayiuse(cpu::x64::avx2)) {
+        roi_align_kernel_.reset(new jit_uni_roi_align_kernel_f32<cpu::x64::avx2>(jcp));
+    } else if (mayiuse(cpu::x64::sse41)) {
+        roi_align_kernel_.reset(new jit_uni_roi_align_kernel_f32<cpu::x64::sse41>(jcp));
+    }
+
+    if (roi_align_kernel_)
+        roi_align_kernel_->create_ker();
+}
+
 void ExperimentalDetectronROIFeatureExtractor::execute(dnnl::stream strm) {
     const int levels_num = inputShapes.size() - INPUT_FEATURES_START;
     const int num_rois = getParentEdgeAt(INPUT_ROIS)->getMemory().getStaticDims()[0];
     const int channels_num = getParentEdgeAt(INPUT_FEATURES_START)->getMemory().getStaticDims()[1];
-    const int feaxels_per_roi = pooled_height_ * pooled_width_ * channels_num;
 
     auto *input_rois = reinterpret_cast<const float *>(getParentEdgeAt(INPUT_ROIS)->getMemoryPtr()->GetPtr());
     auto *output_rois_features = reinterpret_cast<float *>(getChildEdgesAtPort(OUTPUT_ROI_FEATURES)[0]->getMemoryPtr()->GetPtr());
@@ -370,42 +235,138 @@ void ExperimentalDetectronROIFeatureExtractor::execute(dnnl::stream strm) {
         output_rois = reinterpret_cast<float *>(getChildEdgesAtPort(OUTPUT_ROIS)[0]->getMemoryPtr()->GetPtr());
     }
 
-    std::vector<int> level_ids(num_rois, 0);
-    redistribute_rois(input_rois, reinterpret_cast<int *>(&level_ids[0]), num_rois, levels_num);
+    std::vector<std::vector<float>> weights(num_rois);
+    std::vector<std::vector<int>> indices(num_rois);
+    std::vector<int> levels(num_rois);
+    std::vector<LevelParams> level_params(levels_num);
 
-    std::vector<float> reordered_rois(4 * num_rois, 0);
-    std::vector<int> original_rois_mapping(num_rois, 0);
-    reord(input_rois, &level_ids[0], num_rois, 4, &reordered_rois[0], &original_rois_mapping[0]);
-
-    std::vector<int> rois_per_level;
-    split_points(level_ids, rois_per_level, levels_num + 1);
-
-    std::vector<float> output_rois_features_temp(feaxels_per_roi * num_rois, 0);
-    for (int i = 0; i < levels_num; ++i) {
-        const int level_rois_offset = rois_per_level[i];
-        const int level_rois_num = rois_per_level[i + 1] - level_rois_offset;
-        if (level_rois_num > 0) {
-            auto *featuremap = reinterpret_cast<const float *>(getParentEdgeAt(INPUT_FEATURES_START + i)->getMemoryPtr()->GetPtr());
-            const int featuremap_height = getParentEdgeAt(INPUT_FEATURES_START + i)->getMemory().getStaticDims()[2];
-            const int featuremap_width = getParentEdgeAt(INPUT_FEATURES_START + i)->getMemory().getStaticDims()[3];
-            ROIAlignForward_cpu_kernel<float>(feaxels_per_roi * level_rois_num,
-                                              featuremap,
-                                              1.0f / pyramid_scales_[i],
-                                              channels_num,
-                                              featuremap_height,
-                                              featuremap_width,
-                                              pooled_height_,
-                                              pooled_width_,
-                                              sampling_ratio_,
-                                              &reordered_rois[4 * level_rois_offset],
-                                              aligned_,
-                                              &output_rois_features_temp[feaxels_per_roi * level_rois_offset]);
-        }
+    for (size_t i = 0; i < levels_num; ++i) {
+        level_params[i].src = reinterpret_cast<const float *>(getParentEdgeAt(INPUT_FEATURES_START + i)->getMemoryPtr()->GetPtr());
+        level_params[i].height = getParentEdgeAt(INPUT_FEATURES_START + i)->getMemory().getStaticDims()[2];
+        level_params[i].width = getParentEdgeAt(INPUT_FEATURES_START + i)->getMemory().getStaticDims()[3];
     }
 
-    std::vector<int> dummy_mapping(num_rois, 0);
-    reord(&output_rois_features_temp[0], &original_rois_mapping[0], num_rois, feaxels_per_roi,
-            output_rois_features, &dummy_mapping[0]);
+    roiTargetLevels(input_rois, &levels[0], num_rois, levels_num);
+
+    const float offset = aligned_ ? 0.5f : 0.0f;
+
+    parallel_for(num_rois, [&](int n) {
+        const int level = levels[n];
+        if (level >= levels_num) {
+            const size_t output_feature_size = channels_num * pooled_height_ * pooled_width_;
+            memset(&output_rois_features[n * output_feature_size], 0, output_feature_size * sizeof(float));
+            return;
+        }
+
+        // precalc
+
+        const float spatial_scale = 1.0f / pyramid_scales_[level];
+
+        const float x1 = input_rois[4 * n + 0] * spatial_scale - offset;
+        const float y1 = input_rois[4 * n + 1] * spatial_scale - offset;
+        const float x2 = input_rois[4 * n + 2] * spatial_scale - offset;
+        const float y2 = input_rois[4 * n + 3] * spatial_scale - offset;
+
+        // force malformed ROIs to be 1x1
+        const float roi_width = std::max(x2 - x1, 1.f);
+        const float roi_height = std::max(y2 - y1, 1.f);
+
+        const float bin_height = roi_height / pooled_height_;
+        const float bin_width = roi_width / pooled_width_;
+
+        const int sampling_ratio_h = sampling_ratio_ == 0 ? static_cast<int>(ceil(bin_height)) : sampling_ratio_;
+        const int sampling_ratio_w = sampling_ratio_ == 0 ? static_cast<int>(ceil(bin_width)) : sampling_ratio_;
+        const int samples_per_bin = sampling_ratio_h * sampling_ratio_w;
+
+        const float sample_distance_h = bin_height / sampling_ratio_h;
+        const float sample_distance_w = bin_width / sampling_ratio_w;
+
+        pre_calc_for_bilinear_interpolate(
+            n,
+            level_params[level].height,
+            level_params[level].width,
+            pooled_height_,
+            pooled_width_,
+            sampling_ratio_h,
+            sampling_ratio_w,
+            y1,
+            x1,
+            bin_height,
+            bin_width,
+            sample_distance_h,
+            sample_distance_w,
+            weights,
+            indices);
+
+        // pooling
+
+        const float *src_data = level_params[level].src;
+        const int height = level_params[level].height;
+        const int width = level_params[level].width;
+
+        const float samples_per_bin_invert = 1.f / samples_per_bin;
+
+        size_t channel_src_offset = 0;
+        const size_t channel_src_offset_inc = height * width;
+        size_t bin_dst_offset = n * channels_num * pooled_height_ * pooled_width_;
+
+        if (roi_align_kernel_) {
+            auto arg = jit_roi_align_call_args();
+            arg.scale = static_cast<const float*>(&samples_per_bin_invert);
+            arg.num_samples = samples_per_bin;
+
+            for (size_t i = 0; i < channels_num; ++i) {
+                arg.src = static_cast<const void*>(&src_data[channel_src_offset]);
+                channel_src_offset += channel_src_offset_inc;
+
+                size_t param_offset = 0;
+                const size_t param_offset_inc = 4 * samples_per_bin;
+                for (size_t ph = 0; ph < pooled_height_; ++ph) {
+                    for (size_t pw = 0; pw < pooled_width_; ++pw) {
+                        arg.buffer = static_cast<void*>(&indices[n][param_offset]);
+                        arg.weights = static_cast<const float*>(&weights[n][param_offset]);
+                        param_offset += param_offset_inc;
+
+                        arg.dst = static_cast<void*>(&output_rois_features[bin_dst_offset]);
+                        ++bin_dst_offset;
+
+                        (*roi_align_kernel_)(&arg);
+                    }
+                }
+            }
+        } else {
+            for (size_t i = 0; i < channels_num; ++i) {
+                size_t param_offset = 0;
+                for (size_t ph = 0; ph < pooled_height_; ++ph) {
+                    for (size_t pw = 0; pw < pooled_width_; ++pw) {
+                        float pooled_value = 0.f;
+                        for (size_t ind = 0; ind < samples_per_bin; ++ind) {
+                            const float src0 = src_data[channel_src_offset + indices[n][param_offset]];
+                            const float src1 = src_data[channel_src_offset + indices[n][param_offset + 1]];
+                            const float src2 = src_data[channel_src_offset + indices[n][param_offset + 2]];
+                            const float src3 = src_data[channel_src_offset + indices[n][param_offset + 3]];
+
+                            float sample_value =
+                                weights[n][param_offset] * src0 +
+                                weights[n][param_offset + 1] * src1 +
+                                weights[n][param_offset + 2] * src2 +
+                                weights[n][param_offset + 3] * src3;
+
+                            param_offset += 4;
+
+                            pooled_value += sample_value * samples_per_bin_invert;
+                        }
+
+                        output_rois_features[bin_dst_offset] = pooled_value;
+                        ++bin_dst_offset;
+                    }
+                }
+
+                channel_src_offset += channel_src_offset_inc;
+            }
+        }
+    });
+
     if (output_rois != nullptr) {
         cpu_memcpy(output_rois, input_rois, 4 * num_rois * sizeof(float));
     }
