+++ src/plugins/intel_cpu/src/nodes/roi_align.h
@@ -10,59 +10,16 @@
 #include <memory>
 #include <vector>
 #include <dnnl_extension_utils.h>
+#include "kernels/roi_align_kernel.h"
 
 namespace ov {
 namespace intel_cpu {
 namespace node {
 
-enum ROIAlignLayoutType {
-    ncsp,
-    blk,
-    nspc
-};
-
 enum ROIAlignedMode {
     ra_asymmetric,
     ra_half_pixel_for_nn,
     ra_half_pixel
-};
-
-struct jit_roi_align_params {
-    Algorithm alg;
-    InferenceEngine::Precision data_prc;
-    int data_size;
-    ROIAlignLayoutType layout;
-    int pooled_h;
-    int pooled_w;
-};
-
-struct jit_roi_align_call_args {
-    // point to srcData for planar
-    // point to srcData address list for other layouts
-    const void *src;
-    const float *weights;
-    const float *scale;
-    void *buffer;
-    void *dst;
-    size_t num_samples;
-    size_t work_amount;
-    size_t src_stride;
-};
-
-struct jit_uni_roi_align_kernel {
-    void (*ker_)(const jit_roi_align_call_args *);
-
-    void operator()(const jit_roi_align_call_args *args) {
-        assert(ker_);
-        ker_(args);
-    }
-
-    explicit jit_uni_roi_align_kernel(jit_roi_align_params jcp) : ker_(nullptr), jcp_(jcp) {}
-    virtual ~jit_uni_roi_align_kernel() {}
-
-    virtual void create_ker() = 0;
-
-    jit_roi_align_params jcp_;
 };
 
 class ROIAlign : public Node {
